Added ceilometer-polling upstart jobs
The ceilometer-central-agent and ceilometer-agent-compute
services are now completely removed from the liberty release
of ceilometer. The new implementation is to use ceilometer-polling
and call the respective pollsters by namespace.
This patch addes the neccessary upstart configurations for
ceilometer-polling and it also removes the deprecated admin_* options
from the ceilometer config file.
This patch set is also forcing ceilometer to use the Keystone
v2.0 API by adding the versioned path uder [service_credentials]
in the ceilometer config file.

Change-Id: Iba2bb148df4f3ae9d16b44c27ddff0c0d3fd29d3
UpgradeImpact
Closes-Bug: #1519423
Closes-Bug: #1519462