Convert AIO bootstrap from bash to Ansible
This patch converts the AIO bootstrap process to use Ansible
instead of bash scripting. The patch also minimises the options
available to focus the role concerned to just handle an AIO
bootstrap, but gives it just enough flexibility to allow the
use of an external MongoDB database for Ceilometer/Aodh and
for a deployer to specify a secondary disk for the AIO to
consume.

A major change is that the AIO bootstrap process no longer
assumes that it can destroy a secondary boot device. It
requires a device name to be provided. This prevents horrible
surprises.

Backport Note:
While this backport respects the differences between the Master
branch and the Kilo branch, it also brings all the gating and
AIO related scripts into line with each other to ensure that all
methods and settings (where applicable) are the same.

The following differences from the Master patch are noted:
 - neutron_rpc_workers must be 0 for Kilo
 - nova_virt_type must be specifically configured for Kilo
 - the in-tree pip_install role is used (configured in
   tests/ansible.cfg) instead of the new independent role
 - All references to Aodh have been removed
 - The haproxy env.d related config task has been removed as
   this does not apply to Kilo
 - Change I98bcb478e99e11956618b2327399e18471aab534 is included
   (Add reliable OS detection for ansible bootstrap)
 - Change I31f63a2998080c6fedab3af7c2b1bb1a987d8c08 is included
   (Use pure variable name in $(()) statement)
 - Change If3a914650749d72c2eb13b9f1307ef7b4319bd2f is included
   (Remove quotes from subshell call in bash script)
 - Change Iec03c768a9fadf82831af4b90db1907abd399966 is included
   (use fastest mirror available in OpenStack-CI)
 - Change I666f53e01066bf8bff4d28fa012eadae7c958116 is included
   (Bootstrap Host with updated CPU Map for OpenStack-CI)
 - Change Ic20ab65dc68a6f82eb5bf0751a968aa662b29eb5 is included
   (Make instance info output to stdout on failure)
 - Change I28af8a587bde629fd0c81a6c4ba47dd64c2e44db is included
   (Use Ansible AIO playbook for run-aio-build.sh)
 - Change I730c838a67ca03413fbdfa43ef7606d8cf2f9cd0 is included
   (Updating AIO docs for Ansible playbook)

Conflicts:
>------ansible-role-requirements.yml
>------scripts/bootstrap-aio.sh
>------scripts/gate-check-commit.sh
>------scripts/scripts-library.sh

Closes-Bug: #1522957
Implements: blueprint convert-aio-bootstrap-to-ansible
Co-Authored-By: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>
Change-Id: I6028952e7260388873f57db47cc3e08126ecc530
(cherry picked from commit 892c7fe46c68bc36e76f21a6b699349167534f46)