Fix neutron issue w/ l2pop
This change resolves an issue where neutron is not able to bind to
a given port because l2 population is disabled and no vxlan multicast
group has been defined. To resolve this the `neutron_l2_population`
variable is being defined and set to "False" in the os_neutron defaults
and the vxlan multicast group will now contain a default value instead
of an empty string.

The change also removes the neutron_l2_population checks in the tasks
and templates because the variable is now being defined.

Change-Id: Ic2973626d88781bfc67a4275afcf9feffeb63f36
Closes-Bug: #1521793
Co-Authored-by: Ville Vuorinen <ville.vuorinen@multim.fi>
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit 00609ea56e09f5dd0488062b64a03420d296e614)