Tidy up notification_driver implementation across projects
The various os_* roles implement the notification driver in
the config files in a number of different ways.

This commit standardises the implementation.

Some of the decisions that were made here (remove the variables for
the driver altogether, do not configure notifications at all if not
ceilometer) are based on an irc discussion and the outcomes.

The following variables have been removed:
 - glance_ceilometer_notification_driver
 - glance_notification_driver
 - neutron_notification_driver

If these items need to be set differently, the config_override
facility may be used.

UpgradeImpact
Bug: #1522506
Change-Id: I1ac7648a545c35f8f77d79f9cb0d9021bdb9e653
(cherry picked from commit 24488f45c66ed29f8c3c4d370e740af325964508)