Enable multipath usage on all OpenStack hosts
This fix adds the multipath-tools package to all nova
nodes in order to support iscsi-multipath inside nova.
Additionally when using multipath, the stock AIO limit
per process (fs.aio-max-nr) is too low and needs to be
increased to at least 128k.

In order to enable this feature in nova the following override
needs to be configured:

nova_nova_conf_overrides:
  libvirt:
      iscsi_use_multipath: true

This is a backport combining two patches:
 - https://review.openstack.org/256082
 - https://review.openstack.org/257104

DocImpact
Closes-Bug: #1524957
Change-Id: I36da282043ce864c8a1fb69a1ee02e3a273c14e4