Add ceilometer messaging_url for each service...
Without specifying messaging_urls, ceilometer only listens on the vhost
specified in the rabbit configuration directives.
To align with the the default OSA architecture, which isolates each service
into it's own respective rabbitmq vhost, this patchset has added jinja code
in the ceilometer.conf.j2 template to specifiy a messaging_url for each service,
 and for each service, each rabbitmq host. This ensures ceilometer is spawning
the correct number of listeners on the correct vhost and consuming notifications
correctly.

In addition, a few small changes were made in the patch set:

* store_events is now True by default
* rabbitmq_userid and rabbitmq_vhost service variables have been moved to group_vars.

Change-Id: I2bd9cf0ac7ca0492f6b2c56897f77117d2ae730b
Depends-On: I98bcd17dac9a47cc19e0d9ec90176945a6d7e355
Closes-Bug: #1526457