Gate: Bootstrap Host with updated CPU Map
The different cloud providers for OpenStack-CI sometimes make use of
hardware which libvirt does not appropriately map. This results in
gate check failures as the nova scheduler is unable to find a
suitable host on which to build the test VM's.

This patch implements an extra set of tasks into the bootstrap
process which adjusts the libvirt cpu map to resolve this issue.

Implements: blueprint gate-split
Change-Id: I666f53e01066bf8bff4d28fa012eadae7c958116