Fix os_swift_sync role's set_weight
The swift_rings.py file that updates the swift ring files are
failing to update the weight in a builder file. This is caused
by the weight passed into the swift-ring-builder CLI needing
to be a string.

This patch converts the weight to a string when passing it to
swift-ring-builder.

Closes-Bug: #1532204
Change-Id: Iceb3f19eb226fdcfd373fa73d8a5d41b41a239af