Include extra conf file updates when updating SHA's
This patch updates the sources-branch-updater convenience script to
make it also check for additional updates relating to policy.json,
paste.ini, rootwrap.conf, rootwrap filters, etc.

This ensures that the SHA updates are more complete and the updates
require less effort to do.

In addition to the script update there are some changes to other
file names and locations in order to facilitate the file's ability
to do everything correctly. The changes enforce a specific convention
for the names and locations of files it manages.

Change-Id: I99899e61e0527639abc8b3bae0296a891607a31d
(cherry picked from commit 7ef2a9e48fc471ef142ab98acecc92d27f1ec3c4)