Cap django_compressor below 2.0
A recent release of django_compressor (2.0) drops support for Django
versions older than 1.8. However in Kilo, Django is capped to below
version 1.8. This creates a conflict and breaks at least Horizon.
Setting the cap here for django_compressor is the least distruptive way
forward.

This patch introduces a temporary patch which will be in place until
the next SHA update which will pick up this cap from the OpenStack
global requirements update.

Related-Bug: #1532048
Change-Id: I91d32525dbb9324fe4b706fda49af43c611d6bfb