Reduce repo_build console verbosity
Currently the repo_build processes output verbosely to stdout and
also log to a log file independently. This is unnecessary as the
log file output is at maximum verbosity according to
https://pip.pypa.io/en/stable/reference/pip/#file-logging

This patch reduces the console output verbosity so that the gate
check console logs are not as large.

Change-Id: If0f11d1d5a60ede85555e9c0ef134cace94dded6
(cherry picked from commit 61086079f0969fbf8a0e7f8316bd4b8fb53984f8)