Resolve MariaDB/Galera cluster startup/logging issues
This patch ensures that MariaDB is given adequate time to start on a
resources constrained system (180s versus the default of 30s),
ensures that the error log is appropriately populated and also
provides a failback restart in the case where there may be a corrupt
sst directory.

In the handler changes:
 - the environment variable "MYSQLD_STARTUP_TIMEOUT" is now being
   passed into the init script because the defaults are not being
   sourced at the init script runtime.
 - the temporary "sst" directory is cleaned up should the handler
   restart fail. This ensurez that a node is in a clean state if a
   leftover sst directory was on the disk which would cause a node
   to fail to join a cluster or bootstrap.

In the task changes:
 - a new configuration file (part of the mariadb package) is being
   removed which has unforeseen options within it causing no logs
   to be created.
 - a mysql ping check is implemented to verify that the service is
   responding after the restart handler is fired.

This patch is based on:
 - https://review.openstack.org/256016
 - https://review.openstack.org/266265
 - https://review.openstack.org/268707

Closes-Bug: #1532761
Closes-Bug: #1533126
Change-Id: I16af30c660790656fc2d59f9943c172b88098905

Wait for galera to respond after restarts

Add a mysql ping check to verify the service is responding
after a restart handler is fired.

Change-Id: Idfc1e1a1113ab0ffa221e4c0a4cc074df23fe89a
(cherry picked from commit f6fb63f3477e7cdada1a1be8d670755a0e4e6f0b)