Resolve MariaDB/Galera cluster startup/logging issues
This patch ensures that MariaDB is given adequate time to start on a
resources constrained system (180s versus the default of 30s),
ensures that the error log is appropriately populated and also
provides a failback restart in the case where there may be a corrupt
sst directory.

In the handler changes:
 - the environment variable "MYSQLD_STARTUP_TIMEOUT" is now being
   passed into the init script because the defaults are not being
   sourced at the init script runtime.
 - the temporary "sst" directory is cleaned up should the handler
   restart fail. This ensurez that a node is in a clean state if a
   leftover sst directory was on the disk which would cause a node
   to fail to join a cluster or bootstrap.

In the task changes a new configuration file, that is part of the
mariadb package, is being removed which has unforeseen options within
it causing no logs to be created.

The default option "galera_innodb_additional_mem_pool_size" was removed
because its no longer valid within MariaDB10 and we'd never caught that
error message until now.

This patch is based on:
 - https://review.openstack.org/256016
 - https://review.openstack.org/266265

Closes-Bug: #1532761
Closes-Bug: #1533126
Change-Id: I16af30c660790656fc2d59f9943c172b88098905