Fix neutron migrations with optional providers
The venv creation process would include all python packages into the venv
even if the optional provider was not used. This causes neutron migration
issues and creates unneeded table. To resolve this, the py_pkg lookup
plugin has been modified to ensure no "optional" pip packages are included
in the base venv.

The optional string was added to the "plumgrid_pip_packages" array to ensure
the lookup plugin is no longer indexing those packages for inclusion in the
base venv.

Note:
This patch includes a tox.ini configuration change to make it use
test-requirements.txt for the pep8 environment dependencies. This is
essential to allow the lint testing to pass and can be reverted later.

Change-Id: I999ebd52d99f984d7bfa19685fc0fa530c2b32a9
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>