Modify docs to more strongly suggest deploying repo-infra_hosts
This change implements documentation changes discussed during the
weekly meeting on January 14, 2016 on freenode in #openstack-meeting-4

Currently, documentation suggests configuring the user variable
openstack_repo_url: "https://rpc-repo.rackspace.com/"

The repos maintained on https://rpc-repo.rackspace.com/ do not
currently support individual point releases (such as 12.0.3 at the
time of this writing).  Releases are supported by name in that repo,
such as 'liberty', 'master', or 'kilo'.

This change advocates more strongly for the use of deploying
repo-infra_hosts, and in the alternative, asks the user to specify
the user variable openstack_release to avoid problems with point
releases in the repository until a time they are implemented upstream.

Change-Id: I36f8893fcd7cb8d6ddbce06b7974e3b16be91b5f