Make sources-branch-updater checkout the SHA properly
Currently the sources-branch-updater script incorrectly handles the
clone/checkout process and ends up always copying the master branch
files into whichever branch it's trying to update.

This fixes it so that it correctly checks out to the correct SHA
before copying the updated files over.

Backport note:
This patch also removes the tweak to the rootwrap configuration
files as the tweak is only applicable to the venv implementation
in Liberty onwards. It does not apply to the Kilo branch.

Change-Id: I684bc060bf0a4bf4398673b4ca84a2660ee6e1b9
(cherry picked from commit d8acb703016b165de403b07350ed2c041a43a8e0)