Update Liberty SHA's - 18 Jan 2016
Updates all repo SHAs to open up work on 12.0.4

This patch includes a normalisation of file names and updates
of paste, policy and rootwrap configurations.

It also updates tempest.conf.j2 to replace ssh_auth_method with
auth_method, and change auth_method to 'keypair' (configured is no
longer an a valid option).

Change-Id: I45b0ccb00093e6fb5327767b01f2a4543749f2e3