Process git repos in role defaults at low priority
Instead of never processing a git repository defined in a role's
defaults, include it only if that repository is not defined somewhere
else (repo_packages). This allows for independent roles to define git
repositories without necessarily being included in this central
repository, or for the possibility of git repo definitions moving more
exclusively to tagged independent roles.

Change-Id: If303a4cd6fab9129b54879df9aed3fc8e88295ed