Added a variable so pkg-locations to be redefined
This change adds a variable to the repo-build.yml play so that
a deployer can redefine the locations for roles and user_variables.
This is useful for downstream deployers whom may not co-locate their
roles within the OSA deployment repository, the default Ansible role
directory, or within the /etc/openstack_deploy configuration directory.

Change-Id: I4a470536203da2b2da18adf1b436503a77b75ce9
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>