Update Liberty SHAs - 21 Jan 2016
Updates all repo SHAs to open up work on 12.0.5

The global-requirement-pin entry for pip is removed as it is already
catered for in the pip_install role and the repeated entry is not
required.

Some projects name their paste config files with an underscore
instead of a dash. This patch ensures that the source-branch-updater
includes those files too when checking for updates.

The following CVE fix is included:

OSSA-2016-003: Heat denial of service through template-validate
https://security.openstack.org/ossa/OSSA-2016-003.html

Change-Id: I168d1ec9d6a0364d05ae65c3d8e53f1b16d51623