Convert available and minimum disk sizes to bytes.
Previously the check-requirements role would assume GB as the unit
of space when gathering disk sizes from ansible_devices. With
drives larger than 1 TB ansible_devices reports the size as a
float. This converts all disk sizes to bytes for consistency within
the comparisons regardless of the size of the disk.

Change-Id: I07b81a8a35197f73cd338fa02cb7b112df15a012
Closes-Bug: 1536726