Use Swift venv Python interpreter
The use of activate_this to access the venv libraries means that the
system site packages are also included when the scripts swift_rings.py
and swift_rings_check.py are run.

This commit updates the interpreter directive for those scripts, when
Swift is deployed in a venv, to use the venv's interpreter.

Change-Id: I3ff0bf60c9ca695cddc28b05a8e8cc67d5f240e5
Closes-bug: 1538993
(cherry picked from commit a82e4821fda5823541ba83c624f67699559a25bf)