Reduce keystone cache expiration time
This commit removes the keystone_cache_expiration_time variable,
allowing deployers to override using standard config overrides.  The
default cache expiration is 600 seconds, and according to dolphm
bumping this up to 5400 seconds is excessive without documented
reasoning.

Note that this patch is related to [1] but does not need the temporary
disabling of the catalog caching as catalog caching was only introduced
in mitaka.

[1] https://review.openstack.org/#/c/271357/8

UpgradeImpact

Change-Id: I63ca8c653ed6f735056128b2ba248b66833f6981