Add conditional for overlay network settings
A conditional has been added to the ``_local_ip`` settings used in
the ``neutron_local_ip`` which removes the hard requirement for an
overlay network to be set within a deployment. If no overlay network
is set within the deployment the ``local_ip`` will be set to the
value of ``ansible_ssh_host``.

Change-Id: I954a46443b5434b589226c4999e71bd6ba6a5c0b
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>