Docs: encourage users to run apt-get dist-upgrade on target hosts
Ubuntu 14.04 is currently provided in several point releases, most
currently 14.04.3.  It is possible that users may initially deploy
roles on a previous point release of Ubuntu, and then later upgrade
system packages.  Most notably, br_netfilter is implemented in it's
own kernel module in kernel versions >= 3.18.  This results in ansible
fact caches being out of date on subsequent runs, and has caused a few
bugs to be filed along the way.

This patch is to hopefully ensure that users are using the most recent
version of packages when beginning with openstack-ansible.

Change-Id: I1fd4f11fc2e4be4fa4335f63031f7b2bdb312d34
Closes-Bug: #1535075