Add notes about bashate E006 exception
bashate tests fail with 'E006: Line longer than 79 columns'.

As many scripts are deployed as templates and use jinja templating,
this rule is very difficult to achieve. It is still considered a
preference and should be a goal to improve readability, within reason.

Validation of this reasoning will be up to the human reviewers instead
of being forced by the lint tests.

Change-Id: I59cedad79814d51b7dc20610ee2eae38d7a83906