Set hard requirements for pip
This change sets the requirement for pip to be pip>7,<8.
This is being done because the global pip wheel building
producing a built wheel for pip8 which causing downstream
breakage in the most current releases of master and liberty.

Change-Id: Iddcbba94c62dc87dac6fc25efb17b99ee8ec37d6
Closes-Bug: #1541193
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry-picked from commit 4a9b7b848adf24d4b6aeca78e527633348b5c64e)