Docs: Troubleshooting cached Ansible facts issues
Add troubleshooting page describing cached facts and providing a command
to remove the current cached facts in the event they become invalid.

The ansible-playbook option `--flush-cache` is broken for a jsonfile
based cache in Ansible 1.9.x. When OpenStack-Ansible moves to Ansible 2,
that should replace the rm command documented here.

Change-Id: I268a7e7c18cc7ed8986418d2a3232beb214341aa
Related-Bug: #1535075