Documentation for adding roles to OSA
Potential contributors, particularly those coming from other parts
of the OpenStack community, have asked for guidance on how to add a
role to OpenStack-Ansible for a particular service which is not
currently represented. They may or may not be aquainted with Ansible
already. This guide is not intended to teach them how to work with
Ansible but rather to supplement the general Ansible documentation
and provide guidance on what to do with Ansible to build a role for
OpenStack-Ansible.

Closes-Bug: #1535915
Change-Id: I4d03885ad2a91eb3c626ca444afdbadddef7cf31