Updated upgrade process for container restarts
This change updates the upgrade documentation and script to support
a rolling deployment/update of new and old container types. While
this change is not strictly required and could be accomplished by
simply calling the `setup-hosts.yml` play, however its come to light
that in "some" circumstances a statful group of containers may be
restarted, all at the same time, leading to a service interuption of
a stateful service.

Change-Id: I055cdbb117069fde331762f55db46d03069ba829
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>