Updated config parsing for listOPS
This change updates the list parsing for all list options passed through the
config_template module using the INI configuration type. This change is needed
so that when passing a list into a configuration file for use within an
OpenStack service oslo.config properlly handles the lists as a comma seperated
value when rendered.

Closes-Bug: #1543588
Change-Id: If106ef16349a88f3fcd363927bd4b27d11cbbbf3
Depends-On: If8055b3362f1d9d1f274baa447fac094a6a91481
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>