Check local and upstream repo for pip before using pypi
This patch adds an extra source to try and use before failing and
reverting to installing using the isolated mode (thereby installing
from pypi).

This should result in the install task succeeding on the first
attempt before the local repo exists, instead of always failing
and falling back to pypi.

Change-Id: I41821afb35c29984bd6372b6628aa3ca8a2a5715