Skip V-38462 in gate and cap ansible-lint
openstack-infra are creating images containing an apt.d configuration
that sets AllowUnauthenticated to true to support their custom apt
repos which contain unsigned packages.  The presence of this file causes
openstack-ansible-security to fail.  We can either remove the file, or
skip the V-38462 tag.  Skipping the tag in the gate is probably the
better long term solution.

Additionally, ansible-lint 2.3.8 was released today and is now causing
a number of ANSIBLE0008 errors to fail the gate.  This is a temporary
cap to unblock the gate.

Change-Id: I2c78f4e913e99e5fc5d84fba9e28197b73ac8f25