Resolve packaging problems in virtualenv(s)
Because VirtualEnv always assumes you will have the latest and
greatest versions of wheel, pip, and setuptools a couple of changes
need to be done to ensure that the virtualenv command is sourcing
the versions of our packages as we intend in all cases. Currently
our venvs are being populated by pip8 and due to several OS
incompatibilities we're causing ourselves intermitent failures
and other harm by not poperly locking this down.

The change simply forces the virtuenv command to use the release
tag source for all package installs when its preping a new venv
for a given service.

Change-Id: I773c64e2ffb397862bd42ab80a4b4657cdc4809a
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>