Pin wheel version in repo_server role
Currently, wheel 0.29.0 is being pulled in and this is building some
wheels that cannot be installed.  For example,
cffi-1.5.2-cp27-cp27mu-linux_x86_64.whl is being built but pip will
not install this.  Previously, when using 0.26.0, we were building
cffi-1.5.2-cp27-none-linux_x86_64.whl.

After chatting in #pypa, it was recommended to either downgrade wheel,
or upgrade pip >= 8.

Change-Id: Ic7a9d9992a88fcd93bd7c19fb7f5c907a7299b07
Closes-Bug: #1545505