change host_containers group names in inventory
In user config, when choosing host names inspired
from containers group names, this will lead in the code
to have host_containers to be identical to group_containers.
This will eventually lead to unintentially having groups added
to itselves in the inventory which will lead to unending
recursion loop in the dynamic_inventory.py.
Changing the way we construct the name of host_containers
will prevent that.

Change-Id: I5edb6091cb10588099f023f3ae961ed714aea2ce
Closes-Bug: 1512883
Closes-Bug: 1528953