Docs: Explanation of dynamic inventory
This patch provides a common place to document the behavior of the
dynamic inventory scripts provided in tree. It is intentionally
high level, so as to be less influenced by minor changes to the
scripts.

The documentation lives in the developer docs for now, but is also
linked from the install guide as it's relevant to deployers as well.

NOTE: This cannot be backported directly to liberty, since the parameter
to dynamic_inventory.py for the configuration directory is different
(--file in liberty, --config in master)

Change-Id: Ib7e8603dd70cff1d6a29730619c023aadce7b5e1