Implement repo_build IRR & Reduce galera/rabbit/horizon/keystone/repo affinity
This combines two patches in order to unblock the master gate:

Patch 1: Implements IRR for the repo_build role
https://review.openstack.org/#/c/284896/14

Patch 2: Reduce galera/rabbit/horizon/keystone/repo affinity to 1
https://review.openstack.org/#/c/287331/1

This patch reduces the affinity for the above-mentioned containers
to 1 in order to reduce the time the integration test takes to
complete. This is being done as a workaround to allow other patches
through which optimise the gate checks as many checks are currently
failing because they run over the 90 minute check window.

This patch is intended to be reverted before the Mitaka release.

Note that some risk of not catching multi-node code paths is managed
through the functional testing in the independent roles.

Change-Id: Iad0cafb9a873fb2a649cd8cbe430e198863f4b73